Fix twitter showcase; reduce fallback inotify sleep from 10s down to 500ms. (#4869)
* fix twitter showcase

* fix twitter again (thanks pytest!!!)

* reduce fallback inotify sleep from 10s to 500ms

GitOrigin-RevId: 758bc1a60fe9afe6609e414a7f5c2a7876809bb0

+++ showcases/twitter/services/api/app/api.py
@@ -56,12 +56,12 @@ async def read_impact(start: int, end: int) -> dict:
         author_meta.coord_shifted,
         author_meta.total_responses,
         author_meta.total_magic_influence,
-        coalesce(agg.responses_count, 0) as responses_count,
+        coalesce(agg.responses_count, 0)::bigint as responses_count,
         coalesce(agg.mean_sentiment, 0) as mean_sentiment,
         coalesce(agg.magic_influence, 0) as magic_influence,
-        coalesce(agg.far_count, 0) as far_count,
-        coalesce(agg.medium_count, 0) as medium_count,
-        coalesce(agg.close_count, 0) as close_count
+        coalesce(agg.far_count, 0)::bigint as far_count,
+        coalesce(agg.medium_count, 0)::bigint as medium_count,
+        coalesce(agg.close_count, 0)::bigint as close_count
     FROM author_meta
     LEFT JOIN (
         SELECT
